requirements.txt
absl-py==2.1.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
array_record==0.5.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
boto3==1.34.149
botocore==1.34.149
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.30.1
colorlog==6.8.2
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.2
decorator==4.4.2
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring_parser==0.16
egl_probe==1.0.2
embreex==2.17.7.post5
envlogger==1.2
etils==1.7.0
exceptiongroup==1.2.2
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.20.0
filelock==3.13.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.2.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.65.1
gym==0.23.1
gym-notices==0.0.8
gymnasium==0.29.0
h11==0.14.0
h5py==3.11.0
hidapi==0.14.0.post2
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
immutabledict==4.2.0
importlib_metadata==8.2.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isaacsim-app==4.1.0.0
isaacsim-benchmark==4.1.0.0
isaacsim-core==4.1.0.0
isaacsim-extscache-kit==4.1.0.0
isaacsim-extscache-kit-sdk==4.1.0.0
isaacsim-extscache-physics==4.1.0.0
isaacsim-gui==4.1.0.0
isaacsim-kernel==4.1.0.0
isaacsim-replicator==4.1.0.0
isaacsim-rl==4.1.0.0
isaacsim-robot==4.1.0.0
isaacsim-robot-motion==4.1.0.0
isaacsim-sensor==4.1.0.0
isaacsim-storage==4.1.0.0
isaacsim-utils==4.1.0.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lit==18.1.8
lxml==5.2.2
m2t2==1.0.0
manifold3d==2.5.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
meshcat==0.3.2
ml-dtypes==0.4.0
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
numpy-quaternion==2023.0.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia_curobo==0.7.4
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
omni-isaac-lab==0.19.4
omni-isaac-lab_assets==0.1.3
omni-isaac-lab_tasks==0.7.10
omniverse-kit==106.0.1.126909
onnx==1.16.1
open3d==0.18.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
osqp==0.6.7.post1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
plotly==5.22.0
pointnet2_ops==3.0.0
prettytable==3.3.0
proglog==0.1.10
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pybind11==2.13.1
pycollada==0.8
pyglet==1.5.29
Pygments==2.18.0
pyngrok==7.2.0
pyparsing==3.1.2
pyperclip==1.9.0
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qdldl==0.1.7.post4
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rl_games==1.6.1
robomimic==0.3.0
rpds-py==0.19.1
rsl_rl==2.0.2
Rtree==1.3.0
s3transfer==0.10.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
sentry-sdk==2.11.0
setproctitle==1.3.3
setuptools==71.0.4
setuptools-scm==8.1.0
shapely==2.0.5
shortuuid==1.0.13
simple_parsing==0.1.5
six==1.16.0
skrl==1.2.0
smmap==5.0.1
sniffio==1.3.1
stable_baselines3==2.3.2
stack-data==0.6.2
svg.path==6.3
sympy==1.12
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensordict==0.3.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
termcolor==2.4.0
tfds-nightly==4.9.6.dev202408120044
threadpoolctl==3.5.0
tifffile==2024.7.24
toml==0.10.2
tomli==2.0.1
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trimesh==4.4.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
u-msgpack-python==2.8.0
urllib3==2.2.2
usd-core==24.8
vhacdx==0.0.8.post1
wandb==0.12.21
warp-lang==1.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xatlas==0.0.9
xxhash==3.4.1
yourdfpy==0.0.56
zipp==3.19.2
